http://codereview.musicbrainz.org/r/339/
Checked in because no comments after two weeks.

Previously we made use of the fact that the order items were added to an enum was consistent with the order of items returned in the array returned by values() command. But on the Musicbrainz side of things there were variations, i.e ArtistType started at zero, but ReleaseType started at one so we had to start modifying the offsets. So now added getByDbId() getMinDbId() and getMaxDbId() for use by indexers and manglers. Release Status is more complicated because the dbId starts at 100, but the search id starts at 1, so added getBySearchId(), getMinSearchId and getMaxSearchId() as well. Changes allow us to remove various constants and literals in the mangler and indexing code.