Fixes #5420 searching for releases by ASIN ( and script)
It currently doesn't work because the ASIN is not analysed when added to the index, but when searching for ANY field the QueryParser uses the StandardUnaccentAnalyser for ALL fields, and the first thing this analyser does is convert the text to lowercase so a search for B00004Y6O9 becaomes a search for b0004y609, but because it wasn't analysed the searchfield is held in the index as B00004Y6O9 .

Fixed by making ASIN field analyzed, also done for script which has similar problem.
Language is ok despite not being analysed as it is always lowercase anyway
Country works because convert to lowercase when added to index , then when returned in output XML  it converted back to uppercase (but this is a hack)

The better way to do this is to use a PerFieldAnalyzerWrapper so that we can specify KeywordAnalyzer for searching fields that have not been analysed at all such as Musicbrainz Ids, and a CaseInsensitiveKeywordAnalyzer for fields like AmazonId that should be case insensitive this will avoid the unnecessary processing done by StandardUnaccentAnalyzer. This is required anyway for Issues like #5395  but because this is a larger change will implement this as part of NGS.