Added some multithreading support.
All addDocuments request are put on a queue to allow another query to be made to database without waiting
for all added documents to be analysed, queue is serviced by a threads at a maximum of the physical no of cpus.
If the max query outperforms the lucene analysis then analysis will switch to main thread because
the pool queue size cannot be larger than the max number of documents returned from one query.
Will get best results on multiple cpu systems accessing database on another system.