Trac #4827 , JIRA SEARCH-16
Fixed now the three hebrew chars are always converted to their Western equivalents at index and search time in a safe way. Meaning you can use either the Hebrew or Western character. (But as explained in bugtracker if using Western double quote you have to ensure it is escaped correctly so that the query  is a valid search query)