SEARCH-7  Index opening / re-opening
if pass init parameter (with any value) as part http request to the server it will (re)open the index.
Do this after startiung tomcat so ready to go before the real request
Do this after rebulding indexes to get tomcat to see new indexes without having to be restarted.