#SEARCH-21. 3 Index without Norms
Set to no norms for single token fields and non analysed fields because we are not doing any index time doc or field boosts and lengthNorm is calculated as (1.0 / Math.sqrt(numTerms)) i.e 1, so no difference for single term field.

So this change will reduce indexing size and increase indexing speed without any lose of functionality as we are currently using Lucene. (Search time boosting is still enabled if required)