#SEARCH-9
Added support for specifying a different Similarity class for an index.

Improvement to previous fix for SEARCH-9. The previous fix disabled norms this meant a match against one alias within many wasn't disavantaged against  a doc that matched one alias in one.  But unfortunately it meant that matches against artist and alias were biased towards alias with many fields because its tf() would be large but its lengthNorm() would not be correspondingly small. With the fix both these values are normalized so that they are not too large or small wrt to other fields such artist