SEARCH-41
Fixed issue with Release SQL in RecordingIndex, this wasnt picked up by the tests because h2 database seems to be more forgiving then Postgres wrt groups even when in POSTGRES mode. Added a test for a recording linked with two releases and upgraded version of h2 in an attempt to create a test that could cause the failure seen when building index, didnt manage it but adding these anyway as they are improvements. (Must remind myself to always do a test build indexes before checking in, the tests themselves are not enough)