#SEARCH-49 Added a way to build the release index using a simple HashMap, whihc works but requires 2.5GB of memory which is problematic for casual users, but I think would be okay for the real server.0 Added a way to build release index using JCS caching system , which hangs when trying to get the keys back Default is still creation of temporary db table, and readded original method using no temporary tables/caching.
Choice is through the -cacheType option defaults to temporarytable, would have used an enum but could not see how to that with  args4j , but this is probably temporary code and will be dropped when we have established the best solution.

Methods using HashMap or JCS require alot more memory because the resultset they are based on requires about 2 GB to be held in memory, once the resultset has been processed use of a hashmap to hold the releasepuid mapping requires about 800MB. Having got this working I was expecting retrivel of the puids from the hashmap to be much quicker than running queries against the tmp table but doesnt seem to be. Creation of temporary table requires about 600 secs to be spent building the table and index which would be required if created by triggers, and still that temporary table has to be queried for each chunk.

In summary none of these methods has seen been improvements in performance for me, but may have more effect for other setups.

Im now going to put in some extra timings to measure the total time spent on each query/map lookup to see if chunk size effects things dramatically.